tests/requirements.txt
pytest-cov
pytest-xdist
pandas